pean

+++ database/migrations/2019_07_31_042934_create_simpanans_table.php
@@ -16,7 +16,7 @@ class CreateSimpanansTable extends Migration
         Schema::create('simpanans', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->string('nama_simpanan');
-            $table->integer('anggotas_id');
+            $table->unsignedBigInteger('anggotas_id');
             $table->date('tgl_simpanan');
             $table->integer('besar_simpanan');
             $table->timestamps();
